Bug 1341285: Webrtc updated to branch 57 rev 52b6562a10b495; initial pull made Feb 2 2017 14:38 EST Pull updated from 71394677e4dc343ca5c0f996037207a9bd9616c9 to 52b6562a10b495 in late May
UltraBlame original commit: fda2b2655b267852bfd29fe60a3ca0664c8de18b

+++ media/webrtc/trunk/webrtc/base/ratetracker.h
@@ -21,19 +21,19 @@ namespace rtc {
 
 class RateTracker {
  public:
-  RateTracker(uint32_t bucket_milliseconds, size_t bucket_count);
+  RateTracker(int64_t bucket_milliseconds, size_t bucket_count);
   virtual ~RateTracker();
 
   
   
   
-  double ComputeRateForInterval(uint32_t interval_milliseconds) const;
+  double ComputeRateForInterval(int64_t interval_milliseconds) const;
 
   
   
   double ComputeRate() const {
     return ComputeRateForInterval(bucket_milliseconds_ *
-                                  static_cast<uint32_t>(bucket_count_));
+                                  static_cast<int64_t>(bucket_count_));
   }
 
   
@@ -49,19 +49,19 @@ class RateTracker {
 
  protected:
   
-  virtual uint32_t Time() const;
+  virtual int64_t Time() const;
 
  private:
   void EnsureInitialized();
   size_t NextBucketIndex(size_t bucket_index) const;
 
-  const uint32_t bucket_milliseconds_;
+  const int64_t bucket_milliseconds_;
   const size_t bucket_count_;
   size_t* sample_buckets_;
   size_t total_sample_count_;
   size_t current_bucket_;
-  uint32_t bucket_start_time_milliseconds_;
-  uint32_t initialization_time_milliseconds_;
+  int64_t bucket_start_time_milliseconds_;
+  int64_t initialization_time_milliseconds_;
 };
 
 }
